--- a/Kernel/Arch/x86/common/CPU.cpp
+++ b/Kernel/Arch/x86/common/CPU.cpp
@@ -6,8 +6,6 @@
 
 #include <AK/Assertions.h>
 #include <Kernel/Arch/x86/CPU.h>
-#include <Kernel/Arch/x86/Processor.h>
-#include <Kernel/KSyms.h>
 #include <Kernel/Panic.h>
 #include <Kernel/Process.h>