--- a/Kernel/Arch/x86/common/Interrupts.cpp
+++ b/Kernel/Arch/x86/common/Interrupts.cpp
@@ -17,6 +17,7 @@
 #include <Kernel/Random.h>
 #include <Kernel/Sections.h>
 #include <Kernel/Thread.h>
+#include <Kernel/ThreadTracer.h>
 
 #include <LibC/mallocdefs.h>