--- a/Kernel/Arch/x86/i386/ASM_wrapper.cpp
+++ b/Kernel/Arch/x86/i386/ASM_wrapper.cpp
@@ -7,7 +7,6 @@
 #include <AK/Types.h>
 
 #include <Kernel/Arch/x86/ASM_wrapper.h>
-#include <Kernel/Arch/x86/Processor.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {