--- a/Kernel/Graphics/Bochs/GraphicsAdapter.cpp
+++ b/Kernel/Graphics/Bochs/GraphicsAdapter.cpp
@@ -6,7 +6,6 @@
 
 #include <AK/Atomic.h>
 #include <AK/Checked.h>
-#include <AK/Singleton.h>
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>