--- a/Kernel/ProcessGroup.h
+++ b/Kernel/ProcessGroup.h
@@ -9,7 +9,6 @@
 #include <AK/IntrusiveList.h>
 #include <AK/RefCounted.h>
 #include <AK/Weakable.h>
-#include <Kernel/Mutex.h>
 #include <Kernel/SpinLock.h>
 #include <Kernel/UnixTypes.h>