--- a/Kernel/RTC.cpp
+++ b/Kernel/RTC.cpp
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <AK/Assertions.h>
 #include <AK/Time.h>
 #include <Kernel/CMOS.h>
 #include <Kernel/RTC.h>