--- a/Kernel/TimerQueue.cpp
+++ b/Kernel/TimerQueue.cpp
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <AK/NonnullOwnPtr.h>
 #include <AK/Singleton.h>
 #include <AK/Time.h>
 #include <Kernel/Scheduler.h>