--- a/Userland/Applets/DesktopPicker/main.cpp
+++ b/Userland/Applets/DesktopPicker/main.cpp
@@ -6,9 +6,7 @@
 
 #include "DesktopStatusWindow.h"
 #include <LibGUI/Application.h>
-#include <LibGUI/Frame.h>
 #include <LibGUI/Painter.h>
-#include <LibGUI/Window.h>
 #include <LibGUI/WindowManagerServerConnection.h>
 #include <WindowServer/Window.h>
 #include <serenity.h>