--- a/Userland/Applications/DisplaySettings/MonitorSettingsWidget.cpp
+++ b/Userland/Applications/DisplaySettings/MonitorSettingsWidget.cpp
@@ -7,11 +7,9 @@
 
 #include "MonitorSettingsWidget.h"
 #include <Applications/DisplaySettings/MonitorSettingsGML.h>
-#include <LibCore/ConfigFile.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Button.h>
 #include <LibGUI/ComboBox.h>
-#include <LibGUI/Desktop.h>
 #include <LibGUI/ItemListModel.h>
 #include <LibGUI/MessageBox.h>
 #include <LibGUI/RadioButton.h>