--- a/Userland/Applications/HexEditor/main.cpp
+++ b/Userland/Applications/HexEditor/main.cpp
@@ -7,7 +7,6 @@
 #include "HexEditorWidget.h"
 #include <LibGUI/Icon.h>
 #include <LibGUI/Menubar.h>
-#include <LibGfx/Bitmap.h>
 #include <stdio.h>
 #include <unistd.h>