--- a/Userland/Applications/IRCClient/IRCLogBuffer.cpp
+++ b/Userland/Applications/IRCClient/IRCLogBuffer.cpp
@@ -5,12 +5,10 @@
  */
 
 #include "IRCLogBuffer.h"
-#include <AK/StringBuilder.h>
 #include <LibWeb/DOM/DocumentFragment.h>
 #include <LibWeb/DOM/DocumentType.h>
 #include <LibWeb/DOM/ElementFactory.h>
 #include <LibWeb/DOM/Text.h>
-#include <LibWeb/HTML/HTMLBodyElement.h>
 #include <time.h>
 
 NonnullRefPtr<IRCLogBuffer> IRCLogBuffer::create()