--- a/Userland/Applications/PixelPaint/PickerTool.cpp
+++ b/Userland/Applications/PixelPaint/PickerTool.cpp
@@ -7,7 +7,6 @@
 #include "PickerTool.h"
 #include "ImageEditor.h"
 #include "Layer.h"
-#include <LibGfx/Bitmap.h>
 
 namespace PixelPaint {