--- a/Userland/Applications/SoundPlayer/PlaylistWidget.cpp
+++ b/Userland/Applications/SoundPlayer/PlaylistWidget.cpp
@@ -6,7 +6,6 @@
 
 #include "PlaylistWidget.h"
 #include "Player.h"
-#include "SoundPlayerWidgetAdvancedView.h"
 #include <AK/LexicalPath.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/HeaderView.h>