--- a/Userland/Applications/Spreadsheet/Workbook.cpp
+++ b/Userland/Applications/Spreadsheet/Workbook.cpp
@@ -9,18 +9,11 @@
 #include "ImportDialog.h"
 #include "JSIntegration.h"
 #include "Readers/CSV.h"
-#include "Writers/CSV.h"
 #include <AK/ByteBuffer.h>
-#include <AK/JsonArray.h>
-#include <AK/JsonObjectSerializer.h>
-#include <AK/Stream.h>
 #include <LibCore/File.h>
-#include <LibCore/FileStream.h>
 #include <LibCore/MimeData.h>
 #include <LibGUI/TextBox.h>
-#include <LibJS/Parser.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <string.h>
 
 namespace Spreadsheet {