--- a/Userland/Applications/SystemMonitor/MemoryStatsWidget.cpp
+++ b/Userland/Applications/SystemMonitor/MemoryStatsWidget.cpp
@@ -6,16 +6,13 @@
 
 #include "MemoryStatsWidget.h"
 #include "GraphWidget.h"
-#include <AK/ByteBuffer.h>
 #include <AK/JsonObject.h>
 #include <LibCore/File.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Label.h>
 #include <LibGUI/Painter.h>
-#include <LibGfx/Font.h>
 #include <LibGfx/FontDatabase.h>
 #include <LibGfx/StylePainter.h>
-#include <stdio.h>
 #include <stdlib.h>
 
 static MemoryStatsWidget* s_the;