--- a/Userland/DevTools/HackStudio/Git/GitRepo.cpp
+++ b/Userland/DevTools/HackStudio/Git/GitRepo.cpp
@@ -6,8 +6,6 @@
 
 #include "GitRepo.h"
 #include <LibCore/Command.h>
-#include <stdio.h>
-#include <stdlib.h>
 
 namespace HackStudio {