--- a/Userland/DevTools/HackStudio/LanguageClient.cpp
+++ b/Userland/DevTools/HackStudio/LanguageClient.cpp
@@ -10,7 +10,6 @@
 #include "ToDoEntries.h"
 #include <AK/String.h>
 #include <AK/Vector.h>
-#include <DevTools/HackStudio/LanguageServers/LanguageServerEndpoint.h>
 #include <LibGUI/Notification.h>
 
 namespace HackStudio {