--- a/Userland/DevTools/Inspector/RemoteObjectGraphModel.cpp
+++ b/Userland/DevTools/Inspector/RemoteObjectGraphModel.cpp
@@ -7,7 +7,6 @@
 #include "RemoteObjectGraphModel.h"
 #include "RemoteObject.h"
 #include "RemoteProcess.h"
-#include <AK/JsonObject.h>
 #include <AK/JsonValue.h>
 #include <LibGUI/Application.h>
 #include <stdio.h>