--- a/Userland/DevTools/Profiler/TimelineTrack.cpp
+++ b/Userland/DevTools/Profiler/TimelineTrack.cpp
@@ -8,7 +8,6 @@
 #include "Profile.h"
 #include "TimelineView.h"
 #include <LibGUI/Painter.h>
-#include <LibGfx/Font.h>
 #include <LibGfx/Palette.h>
 
 namespace Profiler {