--- a/Userland/DevTools/UserspaceEmulator/SoftCPU.cpp
+++ b/Userland/DevTools/UserspaceEmulator/SoftCPU.cpp
@@ -8,9 +8,7 @@
 #include "SoftCPU.h"
 #include "Emulator.h"
 #include <AK/Assertions.h>
-#include <AK/BitCast.h>
 #include <AK/Debug.h>
-#include <math.h>
 #include <stdio.h>
 #include <string.h>
 #include <unistd.h>