… Tests/AK/TestBase64.cpp
… Tests/AK/TestFind.cpp
… Tests/AK/TestJSON.cpp
… Tests/AK/TestQuickSort.cpp
… Tests/AK/TestUFixedBigInt.cpp
… Userland/Libraries/LibCore/AnonymousBuffer.cpp
… Userland/Libraries/LibCore/Command.cpp
… Userland/Libraries/LibCore/ConfigFile.cpp
… Userland/Libraries/LibCore/DateTime.cpp
… Userland/Libraries/LibCore/EventLoop.cpp
… Userland/Libraries/LibCore/FileWatcher.cpp
… Userland/Libraries/LibCore/Notifier.cpp
… Userland/Libraries/LibCore/Socket.cpp
… Userland/Libraries/LibCore/UDPServer.cpp
… Userland/Libraries/LibGUI/AbstractSlider.cpp
… Userland/Libraries/LibGUI/AbstractTableView.cpp
… Userland/Libraries/LibGUI/Button.cpp
… Userland/Libraries/LibGUI/CommonLocationsProvider.cpp
… Userland/Libraries/LibGUI/Desktop.cpp
… Userland/Libraries/LibGUI/DisplayLink.cpp
… Userland/Libraries/LibGUI/FontPicker.cpp
… Userland/Libraries/LibGUI/GMLParser.cpp
… Userland/Libraries/LibGUI/GMLSyntaxHighlighter.cpp
… Userland/Libraries/LibGUI/INISyntaxHighlighter.cpp
… Userland/Libraries/LibGUI/InputBox.cpp
… Userland/Libraries/LibGUI/Label.cpp
… Userland/Libraries/LibGUI/MessageBox.cpp
… Userland/Libraries/LibGUI/ModelSelection.cpp
… Userland/Libraries/LibGUI/Slider.cpp
… Userland/Libraries/LibGUI/TableView.cpp