Here's some numbers just for fun, the benchmark is very naive and variable.
The test also leans pretty hard on the kernel's disk cache, so these numbers have very little weight...

Here's my system info:
```
$ inxi -C -D -m

Memory:    RAM: total: 31.18 GiB used: 18.82 GiB (60.4%)
           Array-1: capacity: 32 GiB slots: 2 EC: None
           Device-1: ChannelA-DIMM0 size: 16 GiB speed: 2400 MT/s
           Device-2: ChannelB-DIMM0 size: 16 GiB speed: 2400 MT/s

CPU:       Info: Quad Core model: Intel Core i7-8705G bits: 64 type: MT MCP L2 cache: 8 MiB
           Speed: 3700 MHz min/max: 800/4100 MHz Core speeds (MHz): 1: 3700 2: 3700 3: 3700 4: 3700 5: 3700 6: 3700 7: 3700
           8: 3700

Drives:    Local Storage: total: 476.94 GiB used: 714.94 GiB (149.9%)
           ID-1: /dev/nvme0n1 vendor: Samsung model: SSD 970 PRO 512GB size: 476.94 GiB
```

Before:
```
(master) $ hyperfine --prepare 'ninja clean && ccache -C' -r 3 'ninja Kernel/Kernel'
Benchmark #1: ninja Kernel/Kernel
  Time (mean ± σ):     93.158 s ±  7.025 s    [User: 637.220 s, System: 29.089 s]
  Range (min … max):   89.025 s … 101.269 s    3 runs
```
After:
```
(cleanup-more-unused-headers) $ hyperfine --prepare 'ninja clean && ccache -C' -r 3 'ninja Kernel/Kernel'
Benchmark #1: ninja Kernel/Kernel
  Time (mean ± σ):     89.024 s ±  1.090 s    [User: 634.288 s, System: 28.569 s]
  Range (min … max):   88.338 s … 90.281 s    3 runs